requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
install==1.3.5
multidict==6.0.5
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
razorpay==1.4.1
requests==2.31.0
sqlparse==0.4.4
twilio==9.0.5
urllib3==2.2.1
yarl==1.9.4
